moving most functions to more suitable files, major refactoring
